refactor: align HTTP test service naming
