+++ actix-http/src/test/mod.rs
@@ -1,10 +1,10 @@
 //! Various testing helpers for use in internal and app tests.
 
-#[cfg(test)]
-pub(crate) mod services;
 mod test_buffer;
 mod test_request;
 mod test_seq_buffer;
+#[cfg(test)]
+pub(crate) mod test_services;
 
 pub use self::{
     test_buffer::TestBuffer,
